typed dict migration: commander, earley lr0 int64 fix, n00b_dev printf fix
- Commander migrated to typed dicts (n00b_string_t * keys)
- Earley lr0_symbol_of_match updated to int64_t + bit-62 encoding
- n00b_dev token dump uses %lld for 64-bit token IDs
- test_dict updated for typed return values
- 107/107 tests pass

+++ src/tools/n00b_dev.c
@@ -141,8 +141,8 @@ dump_tokens(n00b_token_stream_t *ts)
             tlen = val.u8_bytes;
         }
 
-        printf("%4d  tid=%-10d L%u:C%u  \"%.*s\"\n",
-               index++, tok->tid, tok->line, tok->column,
+        printf("%4d  tid=%-20lld L%u:C%u  \"%.*s\"\n",
+               index++, (long long)tok->tid, tok->line, tok->column,
                (int)tlen, text);
     }
 }
